refactor: button style infile
